Heavily refactored & now just posts she's dead
As the title suggests, I refactored the hell out of this code to add in a lot more formal formatting. 

but also on top of that, I modified the code to no longer use the list of 9 possible statuses like it was before. Personally just because I thought it was getting very repetitive over and over. (How many ways can you say someone is dead in a funny way?). 

Now it will just say "She's Dead." every day like it used to before. Anyway, I'm pretty much done with this project but wanted to give it one final touch up before letting it stay automated and never touching the code again. It's been fun. Time to move onto bigger and better things. 

Truth be told I don't even touch meme bot or nature bot anymore either. But I have nothing to add to them without doing a BUNCH more code to them that I'm not motivated to do anymore. 

Anyway, check out the rest of my repos, I may be working on something cooler. Otherwise, thanks for everything friends. o7